Role REST API raises 404 when role does not exist
git-svn-id: https://svn.alfresco.com/repos/alfresco-open-mirror/alfresco/HEAD@15901 7033f55c-6c29-4ad8-afba-765cb616e738

Former-commit-id: 604932f777cce803b3eadb99505a9694dfd83925

+++ root/modules/dod-5015/source/java/org/alfresco/module/org_alfresco_module_dod5015/script/admin/RmRoleDelete.java
@@ -46,6 +46,7 @@ import org.apache.commons.logging.LogFactory;
  */
 public class RmRoleDelete extends DeclarativeWebScript
 {
+    @SuppressWarnings("unused")
     private static Log logger = LogFactory.getLog(RmRoleDelete.class);
     
     private RecordsManagementService rmService;
@@ -76,6 +77,12 @@ public class RmRoleDelete extends DeclarativeWebScript
         
         List<NodeRef> roots = rmService.getRecordsManagementRoots();
         NodeRef root = roots.get(0);
+     
+        // Check that the role exists
+        if (rmSecurityService.existsRole(root, roleParam) == false)
+        {
+            throw new WebScriptException(Status.STATUS_NOT_FOUND, "The role " + roleParam + " does not exist on the records managment root " + root);
+        }
         
         rmSecurityService.deleteRole(root, roleParam);
